requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
courlan==0.9.4
cron-descriptor==1.4.0
dataclasses-json==0.5.14
dateparser==1.1.8
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
frozenlist==1.4.0
htmldate==1.5.1
idna==3.4
jusText==3.0.0
kombu==5.3.2
langchain==0.0.292
langcodes==3.3.0
langsmith==0.0.37
lxml==4.9.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.25.2
openai==0.28.0
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
regex==2023.8.8
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
tenacity==8.2.3
tld==0.13
tqdm==4.66.1
trafilatura==1.6.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
